requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
Pillow==11.2.1
numpy==1.26.4
opencv-python==4.9.0.80
easyocr==1.7.2
pydantic==2.11.5
llama-cpp-python[server]==0.3.9
torch==2.7.0
jinja2==3.1.3 